+++ core/src/main/scala/com/krux/hyperion/common/SparkCommandRunner.scala
@@ -2,11 +2,11 @@ package com.krux.hyperion.common
 
 import com.krux.hyperion.HyperionContext
 import com.krux.hyperion.adt.HString
-import com.krux.hyperion.resource.{SparkCluster, Resource}
+import com.krux.hyperion.resource.{LegacySparkCluster, Resource}
 
 trait SparkCommandRunner {
 
-  def jobRunner(runsOn: Resource[SparkCluster])(implicit hc: HyperionContext): HString =
+  def jobRunner(runsOn: Resource[LegacySparkCluster])(implicit hc: HyperionContext): HString =
     if (runsOn.asManagedResource.exists(_.releaseLabel.nonEmpty)) {
       // Note calling "spark-submit" directly through common-runner.jar is not supported in the
       // current implementation because SparkStep does not accept local jar and the generated
@@ -16,7 +16,7 @@ trait SparkCommandRunner {
       s"${hc.scriptUri}run-spark-step.sh"
     }
 
-  def scriptRunner(runsOn: Resource[SparkCluster]): HString =
+  def scriptRunner(runsOn: Resource[LegacySparkCluster]): HString =
     // Note we cannot use "command-runner.jar" until SparkStep is reworked to accept local jars and
     // serialization generates the correct arguments
     "s3://elasticmapreduce/libs/script-runner/script-runner.jar"