+++ examples/src/main/scala/com/krux/hyperion/examples/ExampleSpark.scala
@@ -4,11 +4,11 @@ import scala.language.postfixOps
 
 import com.krux.hyperion.Implicits._
 import com.krux.hyperion.action.SnsAlarm
-import com.krux.hyperion.activity.{ SparkActivity, SparkStep, SparkTaskActivity }
+import com.krux.hyperion.activity.{ LegacySparkActivity, SparkStep, SparkTaskActivity }
 import com.krux.hyperion.common.S3Uri
 import com.krux.hyperion.datanode.S3DataNode
 import com.krux.hyperion.expression.{ Format, Parameter, RuntimeNode }
-import com.krux.hyperion.resource.SparkCluster
+import com.krux.hyperion.resource.LegacySparkCluster
 import com.krux.hyperion.{ DataPipelineDef, HyperionCli, HyperionContext, Schedule }
 import com.typesafe.config.ConfigFactory
 
@@ -42,7 +42,7 @@ object ExampleSpark extends DataPipelineDef with HyperionCli {
     .withRole("DataPipelineDefaultResourceRole")
 
   // Resources
-  val sparkCluster = SparkCluster()
+  val sparkCluster = LegacySparkCluster()
     .withTaskInstanceCount(instanceCount)
     .withTaskInstanceType(instanceType)
     .withInitTimeout(5.hours)
@@ -54,7 +54,7 @@ object ExampleSpark extends DataPipelineDef with HyperionCli {
     .withInput(dataNode)
     .withArguments(
       target,
-      Format(SparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd")
+      Format(LegacySparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd")
     )
 
   // Second activity
@@ -62,19 +62,19 @@ object ExampleSpark extends DataPipelineDef with HyperionCli {
     .withMainClass("com.krux.hyperion.ScoreJob1")
     .withArguments(
       target,
-      Format(SparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd"),
+      Format(LegacySparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd"),
       "denormalized"
     )
 
   val scoreStep2 = SparkStep(jar)
     .withMainClass("com.krux.hyperion.ScoreJob2")
-    .withArguments(target, Format(SparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd"))
+    .withArguments(target, Format(LegacySparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd"))
 
   val scoreStep3 = SparkStep(jar)
     .withMainClass("com.krux.hyperion.ScoreJob3")
     .withArguments(
       target,
-      Format(SparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd"),
+      Format(LegacySparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd"),
       "value1,value2"
     )
 
@@ -82,11 +82,11 @@ object ExampleSpark extends DataPipelineDef with HyperionCli {
     .withMainClass("com.krux.hyperion.ScoreJob4")
     .withArguments(
       target,
-      Format(SparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd"),
-      "value1,value2," + Format(SparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd")
+      Format(LegacySparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd"),
+      "value1,value2," + Format(LegacySparkActivity.ScheduledStartTime - 3.days, "yyyy-MM-dd")
     )
 
-  val scoreActivity = SparkActivity(sparkCluster)
+  val scoreActivity = LegacySparkActivity(sparkCluster)
     .named("scoreActivity")
     .withSteps(scoreStep1, scoreStep2, scoreStep3, scoreStep4)
     .onSuccess(mailAction)